refactor: extract PDF region fragments
